Refacto restore from checkpoint

+++ helix-p4d/files/restore.sh
@@ -1,20 +1,20 @@
 #!/bin/bash
 
-## Test for latest link
-if [ ! -L "$P4CKP/latest" ]; then
-	echo "Link not found - looking for checkpoint"
-	/usr/local/bin/latest_checkpoint.sh
-fi
-
 ## Test Checkpoint exists
 if [ ! -L "$P4CKP/latest" ]; then
 	echo "Error: Checkpoint for link $P4CKP/latest not found."
 	exit 255
 fi
 
-## Stop Perforce
-#p4 admin stop
-#until ! p4 info -s 2> /dev/null; do sleep 1; done
+## Check P4CASE environment variable
+if [ -z "${P4CASE:-}" ]; then
+	echo "Error: P4CASE environment variable is not set."
+	exit 255
+fi
+if [ "${P4CASE:-}" != "-C0" ] && [ "${P4CASE:-}" != "-C1" ]; then
+	echo "Error: P4CASE must be set to -C0 (Unix-style) or -C1 (Windows-style)."
+	exit 255
+fi
 
 ## Remove current data base
 rm -rf $P4ROOT/*
@@ -23,13 +23,9 @@ rm -rf $P4ROOT/*
 echo $P4NAME > $P4ROOT/server.id
 
 ## Restore and Upgrade Checkpoint
+echo "Restoring checkpoint..."
 p4d $P4CASE -r $P4ROOT -jr -z $P4CKP/latest
 p4d $P4CASE -r $P4ROOT -xu
 
 ## Set key environment variables
 p4d $P4CASE -r $P4ROOT "-cset security=2"
-p4d $P4CASE -r $P4ROOT "-cset ${P4NAME}#server.depot.root=${P4DEPOTS}"
-p4d $P4CASE -r $P4ROOT "-cset ${P4NAME}#journalPrefix=${P4CKP}/${JNL_PREFIX}"
-
-## Start Perforce
-p4d $P4CASE -r$P4ROOT -p$P4TCP -L$P4LOG -J$P4JOURNAL -d
